refactor(email): split Resend adapter into send, webhook, inbound, and validate modules
Keep packages/email/resend.ts as the public barrel. Put SDK construction in sdk.ts so there is one live adapter.
