+++ packages/email/resend/sdk.ts
@@ -1,0 +1,5 @@
+import { Resend } from "resend";
+
+export function createResendClientInstance(apiKey: string): Resend {
+  return new Resend(apiKey);
+}
